Add OSGI manifest metadata
Added OSGI metadata to the manifest, so the library can be deployed in
OSGI environments. The headers where first generated with the bnd tool,
then manually copied inside the manifest template file with the
appropriate variable substitutions.
Also added a timestamp variable (BUILD_TSTAMP) in the main
CmakeLists.txt for use in the manifest file (and elsewhere, if needed).

Signed-off-by: Alan Alberghini <alan@iooota.com>

[closes #23]