Fix java objects memory leak
Objects created by jni->NewObject inside callback code blocks
are not being released by the JVM as there is still a reference to the object.
Those object references are accumulated on the thread stack and
eventually occupy all the available space in the heap.

Deleting the LocalReference as soon as the object is no longer required allows
the GC to dispose those temporary objects

Signed-off-by: Tzafrir Poupko <tzafrir.poupko@earlysense.com>