Fixed typo in BluetoothDevice.cxx Java_tinyb_BluetoothDevice_getTxPower instead of getTXPower
Fixed class name in helper.cxx get_bluetooth_type, "Ltinyb/BluetoothType"
instead of "l" 

Added a method removeDevices() in BluetoothAdapter class to remove all
the known devices (getDevices() will return an empty list after this call).

Added startNearbyDiscovery() and stopNearbyDiscovery() in BlutoothManager
main class. The start method will clear the internal device list and
send a notification to a listener when a new device is found.
The implementation is a simple thread which will poll each x miliseconds
the list of known devices and compare it to an internal list of known
devices, if the device is new, an event is fired to the passed listener.

Signed-off-by: sbodmer <sbodmer@lsi-media.ch>