Adapter::startDiscovery(): removeDevices() will not work as previously scanned devices may still be in use!
In an multi device environment, it is possible to have one device
already scanned and 'in use' w/ GATT processing
while we enter a new discovery for additional devices.

We need to think about a smart method to kick out 'old devices'
which are not in use and are no more available.

This is curently only possible for:
- end of device operation, manually remove from list
- communication breakup, i.e. loss of [GATT] communication