Fix Whitelist to DBTManager/DBTAdapter and test (Result: Auto-connection but 1+ second slower)
Essential for the whitelist autoconnect is the 'action' parameter,
which is now declared as HCIWhitelistConnectType (BTTypes).
BlueZ Linux kernel sources describe the semantics.

We are now technically enabled to use the auto-connect
and don't even required a discovery phase.
However, the device name is not included in the connect EIR data.

Further, performance lacks heavily similar to no connect
and is more than 1s slower than our procedure below:
- startDiscovery
- device-found
- stopDiscovery
- connect
- GATT operation
...
- startDiscovery