JNI Callback listener issued from a native thread must not mute exceptions (forward to java exception)
All JNI Callback listener which are issued from a native thread don't return to the JVM,
but its native caller thread.

Catching C++ exceptions and forwarding them to the JVM simply mutes the whole exception,
as it will never get checked and processed.

Therefor remove this 'exception muting'
in the JNI AdapterStatusListener and GATTCharacteristicListener instances.

Instead we decorate all listener/callback invocations with exception handling
to properly report this 'user exception' and allow the native thread to continue!

On the native side this has been completed for
AdapterStatusListener, GATTCharacteristicListener and MgmtAdapterEventCallback.
Hence all potential user exceptions covered in the native implementation.