GATTHandler robustness: Fail hard on 'send' and 'sendWithReply': disconnect GATT and device, then throw exception
The hard failure (low tolerance) for GATT/L2CAP communication with one device is mandatory
to allow the user to rewind into a proper device disconnect state and retry later.

The exception is essential for the user to give up GATT communication and end processing,
waiting for the next discovery or connect.

The device disconnect is therefor important to get back into a clean state
and have the user consume the DISCONNECT event.