L2CAPComm: connect/disconnect: Don't fail if connect and _dd doesn't match -> intended
It is intended to allow e.g. disconnect from multiple threads at the 'same time',
hence we utilize the 'atomic<bool> connect' to bail out early if already called.

In this case and if _dd (l2cap device descriptor) is still valid,
a concurrent thread is still about to close the _dd.