DBTAdapter: Have non-enabled adapter to be valid, use lazy HCI initialization; DBTManager: Validate enabled for non-user default adapter.
DBTAdapter: Have non-enabled adapter to be valid, use lazy HCI initialization

In a multi adapter setting, it is possible to have one adapter not being enabled,
i.e. blocked from operations (e.g. rfkill).
HCI communication is not possible in such cases and 'openHCI()' will fail.

We still need to treat such adapter as valid, as they are,
and distinguish the states: 'valid' and 'enabled'.

This allows 'delivery' of all adapters to a manager
and allows querying whether they are enabled.

Therefor, the native DBTAdapter will not issue an aggregated 'openHCI()'
at construction and initial validation, but treat the HCI instance as a singleton
to be lazily created via 'getHCI()' if not yet existing.
Here the required listener are also attached to the instance.

In the future, this shall also allow dynamically adding/removing
and enabling/disabling adapters.

Hence before using a chosen adapter, one should query its enabled state
via 'isEnabled()', which will also potentially initializing HCI.

+++

DBTManager: Validate enabled for non-user default adapter.

On the Java side, we can chose a default adapter via the property 'org.tinyb.default_adapter'.
If so chosen, it shall be the default adapter, regardless of its 'enabled' state.

Otherwise, the default adapter shall be the first enabled adapter in the list.