HCIHandler::disconnect(..): Resolve CONNECTION_ALREADY_EXISTS re-connect failure on CSR adapter post lost connection (ioError)
Always issue DISCONNECT command, even in case of an ioError (lost-connection),
see Issue #124 fast re-connect on CSR adapter.
This will always notify the adapter of a disconnected device.

We still send the DISCONN_COMPLETE in such case,
i.e. not waiting for the lagging DISCONN_COMPLETE event from the adapter.