HCIHandler: Unify processSimpleCommand/processStructCommand -> processCommandStatus/processCommandComplete using base HCICommand
Whether a simple HCICommand or HCIStructCommand<..> is being sent is up to the caller
and agnostic to the methods.

We either process a command with intermediate status respone (lagging result)
and hence no detailed result structure/data

- or -

we process a command with complete status (and completed operation)
potentially with detailed result structure/data.