DBTEnv: Use plain 'foo.bar' names, not 'foo_bar'; Add 'exploding' name values, i.e. "direct_bt.debug" = "hci.event,manager.event"
The renaming of 'foo.bar' to 'foo_bar' is not necessary and only complicates things, hence dropped.

+++

Adding 'Exploding variable-name values'
=======================================
If the value of variable 'direct_bt.debug' is neither 'true' or 'false',
it is treated as a list of sub-variable names separated by comma ','.

Each sub-variable name will be trimmed and if not zero-length
appended to the basename "direct_bt.debug" with a dot '.'.

The new variable name will be set in the environment with value 'true'.

The boolean variable value having an exploded value is considered true!

This is supported for DEBUG 'direct_bt.debug' and VERBOSE 'direct_bt.verbose'.

Example:
  "direct_bt.debug" = "hci.event,manager.event,gatt.data".

This leads DBTEnv to set the following environment variables:
  "direct_bt.debug.hci.event"  = "true"
  "direct_bt.debug.manager.event"  = "true"
  "direct_bt.debug.gatt.data" = "true"

In this case, DEBUG "direct_bt.debug" is also considered true!

+++