PlatformToolkit: Properly map 'os.arch' -> CPUType (reduction) -> 'os_arch' (our common name)
As it turned out that OpenJDK uses 'arm' for 'os.arch' on an armhf machine,
we are required to perform some mapping, i.e. reduction of 'os.arch' and map.

Hence the reuse of CPUFamily, CPUType (reduction) and ABIType
which eventually map to our common name.

This is also required as we already have chosen our common name,
based on the 'archabi' name used on the platform.
Too bad OpenJDK couldn't use same name here,
so they use 'armhf' for the native installation folder
but 'arm' for 'os.arch'.

However, introducing this 'reduce map' will easy further support.

Tested on OS Linux
- i386
- amd64
- armhf
- aarch64 (not yet, but names are known)