BluetoothGattCharacteristic: API Change: 'writeValue(byte[] value)' -> 'writeValue(byte[] value, boolean withResponse)'
Add both writeValue(..) mapping to Java
[1] BT Core Spec v5.2: Vol 3, Part G GATT: 4.9.1 Write Characteristic Value Without Response
[2] BT Core Spec v5.2: Vol 3, Part G GATT: 4.9.3 Write Characteristic Value

Previously only [1] existed and [2] seemingly isn't supported by DBus yet(?)

Adding the additional paramter 'boolean withResponse' to determine which variant is desired,
while removing the original method type.

The TinyB/DBus implementation will throw a BluetoothException if 'withResponse' == true,
as it is not supported.

This change allows clarification of the actual method being desired
and supports a fail fast if not supported.